+++ ESP32/.vscode/extensions.json
