+++ ESP32/platformio.ini
@@ -13,4 +13,5 @@ platform = espressif32
 board = esp32doit-devkit-v1
 framework = arduino
 upload_port = COM7
-monitor_speed=115200
+monitor_speed = 115200
+lib_deps = arduino-libraries/Arduino_JSON@^0.1.0
