+++ ESP32/src/main.cpp
@@ -1,0 +1,183 @@
+#include <WiFi.h>
+#include <HTTPClient.h>
+#include <Arduino_JSON.h>
+#include <Arduino.h>
+
+const char *ssid = "Lolomolo";
+const char *password = "21plkoyr92";
+
+//Your Domain name with URL path or IP address with path
+const char *serverName = "http://10.0.0.4:3000/api/data";
+
+// the following variables are unsigned longs because the time, measured in
+// milliseconds, will quickly become a bigger number than can be stored in an int.
+unsigned long lastTime = 0;
+// Timer set to 10 minutes (600000)
+//unsigned long timerDelay = 600000;
+// Set timer to 5 seconds (5000)
+unsigned long timerDelay = 15000;
+
+void setup()
+{
+  Serial.begin(115200);
+
+  WiFi.begin(ssid, password);
+  Serial.println("Connecting");
+  while (WiFi.status() != WL_CONNECTED)
+  {
+    delay(500);
+    Serial.print(".");
+  }
+  Serial.println("");
+  Serial.print("Connected to WiFi network with IP Address: ");
+  Serial.println(WiFi.localIP());
+
+  Serial.println("Timer set to 5 seconds (timerDelay variable), it will take 5 seconds before publishing the first reading.");
+
+}
+
+
+float readTemp(){
+  float reading = 0;
+  for(int i =0; i < 1024; i++){
+    reading += analogRead(34);
+  }
+  float avgReading = reading / 1024;
+  Serial.println(avgReading);
+  float volts = avgReading*3.3 / 4096;
+  Serial.println(volts);
+  float temp = (volts-.5)/0.0195;
+
+  Serial.println(temp);
+  return temp;
+}
+void loop()
+{
+
+  //Send an HTTP POST request every 10 minutes
+  if ((millis() - lastTime) > timerDelay)
+  {
+    //Check WiFi connection status
+    if (WiFi.status() == WL_CONNECTED)
+    {
+      WiFiClient client;
+      HTTPClient http;
+
+      // Your Domain name with URL path or IP address with path
+      http.begin(client, serverName);
+
+      // Specify content-type header
+      // http.addHeader("Content-Type", "application/x-www-form-urlencoded");
+      // Data to send with HTTP POST
+      //String httpRequestData = "UUID=fa0c33ce-68ab-44ab-9ac8-16de4f47cd0a sensorType=Temperature sensorValue=30";
+      // Send HTTP POST request
+      //int httpResponseCode = http.POST(httpRequestData);
+
+      // If you need an HTTP request with a content type: application/json, use the following:
+      http.addHeader("Content-Type", "application/json");
+      float temp = readTemp();
+      String httpRequestData = "{\"UUID\":\"fa0c33ce-68ab-44ab-9ac8-16de4f47cd0a\",\"sensorType\":\"Temperature\",\"sensorValue\":\"" + String(temp, DEC) + "\"}";
+      int httpResponseCode = http.POST(httpRequestData);
+
+      // If you need an HTTP request with a content type: text/plain
+      //http.addHeader("Content-Type", "text/plain");
+      //int httpResponseCode = http.POST("Hello, World!");
+
+      Serial.print("HTTP Response code: ");
+      Serial.println(httpResponseCode);
+
+      // Free resources
+      http.end();
+    }
+    else
+    {
+      Serial.println("WiFi Disconnected");
+    }
+    lastTime = millis();
+  }
+}
+
+/*
+#include <WiFi.h>
+#include <HTTPClient.h>
+#include <Arduino.h>
+#include <Arduino_JSON.h>
+
+const char *ssid = "Lolomolo";
+const char *password = "21plkoyr92";
+
+//Your Domain name with URL path or IP address with path
+String serverName = "http://10.0.0.4:3000/api/data";
+
+// the following variables are unsigned longs because the time, measured in
+// milliseconds, will quickly become a bigger number than can be stored in an int.
+unsigned long lastTime = 0;
+// Timer set to 10 minutes (600000)
+//unsigned long timerDelay = 600000;
+// Set timer to 5 seconds (5000)
+unsigned long timerDelay = 5000;
+
+String sensorReadings;
+float sensorReadingsArr[3];
+
+void setup()
+{
+  Serial.begin(115200);
+
+  WiFi.begin(ssid, password);
+  Serial.println("Connecting");
+  while (WiFi.status() != WL_CONNECTED)
+  {
+    delay(500);
+    Serial.print(".");
+  }
+  Serial.println("");
+  Serial.print("Connected to WiFi network with IP Address: ");
+  Serial.println(WiFi.localIP());
+
+  Serial.println("Timer set to 5 seconds (timerDelay variable), it will take 5 seconds before publishing the first reading.");
+}
+
+
+
+void loop()
+{
+  //Send an HTTP POST request every 10 minutes
+  if ((millis() - lastTime) > timerDelay)
+  {
+    //Check WiFi connection status
+    if (WiFi.status() == WL_CONNECTED)
+    {
+      HTTPClient http;
+
+      String serverPath = serverName + "?temperature=24.37";
+
+      // Your Domain name with URL path or IP address with path
+      http.begin(serverPath.c_str());
+
+      // Send HTTP GET request
+      int httpResponseCode = http.GET();
+
+      if (httpResponseCode > 0)
+      {
+        Serial.print("HTTP Response code: ");
+        Serial.println(httpResponseCode);
+        String payload = http.getString();
+        Serial.println(payload);
+      }
+      else
+      {
+        Serial.print("Error code: ");
+        Serial.println(httpResponseCode);
+      }
+      // Free resources
+      http.end();
+    }
+    else
+    {
+      Serial.println("WiFi Disconnected");
+    }
+    lastTime = millis();
+  }
+}
+*/
